fix: add new agencyName and create_date fields to AskerInfoComponent, revert AskerInfoData
