+++ src/extensions/resources/i18n/overwrites/de/common.json
@@ -247,11 +247,13 @@
 	},
 	"profile": {
 		"age": "Alter",
+		"agencyName": "Beratungsstelle",
 		"counsellingRelation": {
 			"parental_counselling": "Elternteil",
 			"relative_counselling": "Angehörige_r",
 			"self_counselling": "Betroffene_r"
 		},
+		"createDate": "Erstellungsdatum",
 		"data": {
 			"emailInfo": "Die Angabe Ihrer E-Mail ist freiwillig und wird ausschließlich verwendet, um Sie über neue Antworten Ihrer Beraterin / Ihres Beraters zu informieren. Ihre E-Mail-Adresse ist für Berater_innen nicht sichtbar."
 		},
