+++ src/extensions/resources/i18n/overwrites/en/common.json
@@ -1,5 +1,7 @@
 {
 	"profile": {
+		"agencyName": "Agency Name",
+		"createDate": "Creation Date",
 		"tools": {
 			"documentation": "Documentation"
 		}
